Match ckan file store logic (#17)
* Match ckan file store logic

ckan file store logic does not care about the filename provided by the user, it also stores
files without filenames nor filetype but as a uuid string in a resource folder.

S3 Plugin was built to store in a resource folder and the filename was used
If not provided it crashed, which was recently corrected.

Only problem is that people who use ckan have gotten use to being able to change
filenames and have the latest file served. Also with deep linking this also points to
old data which users do not like.